Move dns server module
